+++ CHANGELOG.md
@@ -82,6 +82,7 @@
 - In the parsers package, create_parser() was moved to __init__ and the factory module was removed.
 - is_proper_concept() renamed to has_proper_concept() (meaning.concepts module).
 - find_concept() renamed to strip_concept() (meaning.concepts module).
+- Improved notebooks visualizations (show(), blocks(), vblocks()).
 
 ### Removed
 - graphbrain.funs module.